backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM SDK - Ollama
langchain-ollama>=0.1.0
langchain-core>=0.3.0
ollama>=0.4.0

# HTTP 客户端
httpx==0.27.0
requests==2.31.0

# WebSocket
websockets==12.0

# 环境变量管理
python-dotenv==1.0.0

# 日志
python-json-logger==2.0.7

# 工具
python-dateutil==2.8.2

# 语音识别 (FunASR)
funasr>=1.0.0
modelscope>=1.10.0
torch>=2.0.0
soundfile>=0.12.1
numpy>=1.24.0

# 音频处理
pydub>=0.25.1
